Chess-AI/requirements.txt
# pygame
python-chess
flask
flask-cors
stockfish
gunicorn
